widgetAdapter<->Dashboard connector fixed, LightRGB and Socket are ready, Processing experimental code to set fixed width, Configure and ModuleEditRequest fixed to interact correctly, some color changes

+++ app/src/main/java/com/lastutf445/home2/loaders/ModulesLoader.java
@@ -530,7 +530,7 @@ public class ModulesLoader {
                     0
             );
 
-            lastSync = DataLoader.getInt("lastSyncModules", 0);
+            lastSync = DataLoader.getLong("lastSyncModules", 0);
             group = Sync.SYNC_MODULES_STATE;
         }
